% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_foci.R
\name{generate_foci}
\alias{generate_foci}
\title{generate_foci}
\usage{
generate_foci(nctids, username, password)
}
\arguments{
\item{nctids}{(character vector): A vector of NCT IDs for which trial foci need to be generated.}

\item{username}{(character): The username for database authentication.}

\item{password}{(character): The password for database authentication.}
}
\description{
This function retrieves trial foci based on a list of NCT IDs from a PostgreSQL database.
It connects to the AACT database, retrieves the major mesh headings associated with the NCT IDs, and
generates major therapeutic focus for each trial according to NLM 2023 MeSH headings.The output is
a data frame containing the NCT IDs and their respective therapeutic focuses.
}
\note{
An account in AACT database is required for credentials
(visit this link for registration : https://aact.ctti-clinicaltrials.org/users/sign_up)
}
