% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foci_euctr.R
\name{get_foci_euctr}
\alias{get_foci_euctr}
\title{get_foci_euctr}
\usage{
get_foci_euctr(trn)
}
\arguments{
\item{trn}{A character string containing a well-formed EUCTR
identifier}
}
\value{
A character string with the contents of the
    "Medical condition field" from EUCTR for the record that
    corresponds to the provided identifier
}
\description{
Downloads the medical condition field from EUCTR for a
    given record identifier
}
\examples{
get_foci_euctr("2010-023457-11")
## "advanced/recurrent ovarian and endometrial cancer"
}
