% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mesh_tree_disease.R
\name{process MeSH tree 2023}
\alias{process MeSH tree 2023}
\title{process MeSH tree 2023}
\arguments{
\item{mtrees}{MeSH Files in ASCII UTF-8 format for year 2023}
}
\value{
mesh_tree containing mesh headings and tree
}
\description{
Read and Process Mesh Tree Data

The code reads an ASCII format UTF-8 Mesh Tree file and processes it to extract relevant information.
It filters the data for tree numbers related to diseases, assigns major tree numbers based on grouping,
and assigns major mesh headings. The resulting clean table of Mesh Tree data can be used as a reference
to assign therapeutic foci for clinical trials on ClinicalTrials.gov with MeSH terms stored in AACT database.
}
\note{
MeSH tree files in a ASCII format UTF-8 and generated every year. Use updated MesH tree
file (link:https://www.nlm.nih.gov/databases/download/mesh.html)
}
